% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\docType{class}
\name{ASP-class}
\alias{ASP-class}
\alias{ASP}
\title{Class \code{ASP}}
\description{
All information stored in ASP object.
   You can use \code{creatASP} to create an ASP object.
   In this package, most of the functions will use
   ASP object as input, and return a modified ASP as well.
}
\section{Slots}{

\describe{
\item{\code{sampletable}}{data.frame. An experiment design table and
these columns are mandatory:
\itemize{
    \item samples
    \item conditions
    \item files_bam
    \item dirs_salmon
    \item read_lengths
    \item library_types
    \item strandedness
}}

\item{\code{tx2gene}}{data.frame. At least three types of
information for each transcript should be provided:
\itemize{
    \item gene_id
    \item gene_symbol
    \item transcript_id
}}

\item{\code{tx2gene_convert}}{data.frame. Automatically generated from tx2gene.}

\item{\code{gtf}}{character. The path to your gtf file.}

\item{\code{gff}}{character. The path to your gff3 file.}

\item{\code{fa}}{character. The path to your genome fasta file.}

\item{\code{fa_transcript}}{character. The path to your transcriptome fasta file.}

\item{\code{genome_version}}{character. hg38 or hg19.}

\item{\code{dir_out}}{character. The parent directory for all the next procedures.}

\item{\code{nproc}}{integer. The number of threads.}

\item{\code{novel}}{boolean. If detect novel splicing or not.}

\item{\code{write_log}}{boolean. If write the stdout and stderr to a log file
beginning with an underscore.}

\item{\code{parallel}}{boolean. If run different tools or multiple samples
in parallel.}

\item{\code{np}}{integer. How many samples will be processed at the same time.}

\item{\code{conda_path}}{character. The path to your conda or vritual environment.
Different tools can rely on different paths.}

\item{\code{basics}}{list. Automatically generated based on the sampletable.}

\item{\code{cmds}}{list. Storing commands of different tools to run or to check.}

\item{\code{results}}{list. Storing results outputted by different tools.}

\item{\code{sig_results}}{list. Results fulfilling significant thresholds.}

\item{\code{parameters}}{list.}

\item{\code{intersection}}{list. Results intersections.}

\item{\code{txdb}}{TxDb. Generated from the gtf file.
\code{\link[GenomicFeatures]{makeTxDbFromGFF}}.}

\item{\code{annotation_db}}{OrgDb. Bioconductor annotation data package.
for example: \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}.}
}}

