% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASP.R
\docType{package}
\name{ASP-package}
\alias{ASP-package}
\title{ASP: A package to facilitate performing differential alternative splicing
    analysis.}
\description{
ASP: A package to facilitate performing differential alternative
    splicing analysis.
}
