% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{bandits}
\alias{bandits}
\title{BANDITS}
\usage{
bandits(
  dir_out,
  sampletable,
  basics,
  nproc,
  conda_path,
  conda_env,
  write_log,
  tx2gene,
  infer_prior = T,
  max_genes_per_group = 50,
  libpaths = "",
  ...
)
}
\arguments{
\item{dir_out, sampletable, basics, nproc, conda_path, conda_env, write_log, tx2gene}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{infer_prior}{Default is TRUE.}

\item{max_genes_per_group}{Default is 50.}

\item{libpaths}{Default is "". If not NULL, should be like this:
c("~/doc/Rlibrary/", "~/doc/Rlibrary4/")}

\item{...}{nothing}
}
\value{
A list containing BANDITS workflow.
}
\description{
Generating commands of BANDITS workflow.
}
\references{
\url{https://bioconductor.org/packages/release/bioc/html/BANDITS.html}.
}
