% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{cash}
\alias{cash}
\title{CASH}
\usage{
cash(
  dir_out,
  basics,
  gtf,
  novel,
  conda_path,
  conda_env,
  write_log,
  cash_jar,
  MergePval = "G",
  ram = "100g",
  ...
)
}
\arguments{
\item{dir_out, basics, gtf, novel, conda_path, conda_env, write_log}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{cash_jar}{No default. The path to cash.jar.}

\item{MergePval}{Default is "G".}

\item{ram}{Default is "100g".}

\item{...}{nothing}
}
\value{
A list containing CASH workflow.
}
\description{
Generating commands of CASH workflow.
}
\references{
\url{https://soft.novelbio.com/cash/}.
}
