% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{create_asp}
\alias{create_asp}
\title{Create an \code{\linkS4class{ASP}} object}
\usage{
create_asp(...)
}
\arguments{
\item{...}{paramters pass to \code{\linkS4class{ASP}}.
These flags are mandatory for running AS tools:
sampletable, gtf, gff, and fa.}
}
\value{
An \code{\linkS4class{ASP}} object.
}
\description{
This function is about how to build an \code{\linkS4class{ASP}}
   object. An \code{\linkS4class{ASP}} object is the base for the whole
   analysing workflow.
}
\examples{
\dontrun{
bams_s1 <- system.file(
  "extdata", glue("star/sample{1:4}/sample{1:4}.SortedByCoord.bam"),
  package = "asp"
)
bams_s2 <- system.file(
  "extdata", glue("star/sample{5:8}/sample{5:8}.SortedByCoord.bam"),
  package = "asp"
)
salmons_s1 <- system.file(
  "extdata", glue("salmon/sample{1:4}/"), package = "asp"
)
salmons_s2 <- system.file(
  "extdata", glue("salmon/sample{5:8}/"), package = "asp"
)
sampleTable <- data.frame(
  samples = glue("sample{1:8}"),
  conditions = rep(c("s1", "s2"), each = 4),
  files_bam = c(bams_s1, bams_s2),
  dirs_salmon = c(salmons_s1, salmons_s2),
  read_lengths = rep(101, 8),
  library_types = rep("paired-end", 8),
  strandedness = rep("no", 8)
)

asp <- create_asp(
  sampletable = sampleTable,
  gtf = "~/doc/reference/gtf/gencode.v32.annotation.gtf",
  gff = "~/doc/reference/gtf/gff/gencode.v32.annotation.gff3",
  fa = "~/doc/reference/fa/GRCh38.p13.genome.fa",
  conda_path = "~/bin/anaconda3/"
)
}
}
