% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{createASP}
\alias{createASP}
\title{Create an \code{\linkS4class{ASP}} object}
\usage{
createASP(...)
}
\arguments{
\item{...}{paramters pass to \code{\linkS4class{ASP}}.}
}
\value{
An \code{\linkS4class{ASP}} object.
}
\description{
This function is about how to build an \code{\linkS4class{ASP}}
   object. An \code{\linkS4class{ASP}} object is the base for the whole
   analysing workflow.
}
