% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{darts}
\alias{darts}
\title{DARTS}
\usage{
darts(
  dir_out,
  basics,
  nproc,
  conda_path,
  conda_env,
  write_log,
  parallel,
  tx2gene,
  rmats_dir = NA,
  cstat = 0.05,
  ...
)
}
\arguments{
\item{dir_out, basics, parallel, tx2gene, nproc, conda_path, conda_env, write_log}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{rmats_dir}{rMATS output directory.}

\item{cstat}{Default is 0.05.}

\item{...}{nothing}
}
\value{
A list containing DARTS workflow.
}
\description{
Generating commands of DARTS workflow.
}
\references{
\url{https://github.com/Xinglab/DARTS}.
}
