% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{intersect_asp}
\alias{intersect_asp}
\title{Intersect the results}
\usage{
intersect_asp(
  asp,
  tools = "all",
  type = "all",
  dir_plot = NULL,
  width = 10,
  height = 7
)
}
\arguments{
\item{asp}{An \code{\linkS4class{ASP}} object.}

\item{type}{One of "SE", "MXE", "A3SS", "A5SS", "RI", "SME", "ALE" and "AFE".
Or use "all" to retrieve the intersections of all AS types.}

\item{dir_plot}{If it's NULL, pdf files will be deposited under the
\emph{dir_out} folder stored in \code{\linkS4class{ASP}}.}

\item{width}{pdf width.}

\item{height}{pdf height.}

\item{tool}{Any combination of "rmats", "cash", "leafcutter", "majiq",
"suppa", "bandits", "psichomics", and "spladder".
Or use "all" to retrieve the intersections of all significant results
stored in your \code{\linkS4class{ASP}} object.}
}
\value{
An \code{\linkS4class{ASP}} object with intersection tibble
    stored in the intersection slot.
}
\description{
This function is to retrieve intersections of significant
    results of different workflows and to visualize it.
}
