% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{jum}
\alias{jum}
\title{JUM}
\usage{
jum(
  dir_out,
  sampletable,
  basics,
  nproc,
  conda_path,
  conda_env,
  write_log,
  REF,
  jum_dir,
  JuncThreshold = 3,
  IRthreshold = 5,
  Condition1_fileNum_threshold = NA,
  Condition2_fileNum_threshold = NA,
  ...
)
}
\arguments{
\item{dir_out, sampletable, basics, nproc, conda_path, conda_env, write_log}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{REF}{An annoation of GenePred format.}

\item{jum_dir}{The driectory containing JUM scripts.}

\item{JuncThreshold}{Please consult JUM website.}

\item{IRthreshold}{Please consult JUM website.}

\item{Condition1_fileNum_threshold}{Default is half of the samples belonging to condition 1.}

\item{Condition2_fileNum_threshold}{Default is half of the samples belonging to condition 2.}

\item{...}{nothing}
}
\value{
A list containing JUM workflow.
}
\description{
Generating commands of JUM workflow.
}
\references{
\url{https://github.com/qqwang-berkeley/JUM}.
}
