% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{leafcutter}
\alias{leafcutter}
\title{LeafCutter}
\usage{
leafcutter(
  dir_out,
  sampletable,
  nproc,
  np,
  gtf,
  conda_path,
  conda_env,
  write_log,
  parallel,
  exon_file = NULL,
  leafcutter_dir,
  ...
)
}
\arguments{
\item{dir_out, sampletable, nproc, np, gtf, conda_path, conda_env, write_log, parallel}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{exon_file}{Any path is okay but should ends with .gz.
Keep NULL will make the file be generated as
\emph{dir_out/LeafCutter/exon_file.gz}.}

\item{leafcutter_dir}{No default. The path to
your downloaded LeafCutter directory.}

\item{...}{nothing}
}
\value{
A list containing LeafCutter workflow.
}
\description{
Generating commands of LeafCutter workflow.
}
\references{
\url{https://davidaknowles.github.io/leafcutter/}.
}
