% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{majiq}
\alias{majiq}
\title{MAJIQ}
\usage{
majiq(
  dir_out,
  basics,
  sampletable,
  nproc,
  gff,
  novel,
  fa,
  conda_path,
  conda_env,
  write_log,
  show_all = F,
  threshold = 0.2,
  ...
)
}
\arguments{
\item{dir_out, basics, sampletable, nproc, gff, novel, fa, conda_path, conda_env, write_log}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{show_all}{Default is FALSE.}

\item{threshold}{Default is 0.2.}

\item{...}{nothing}
}
\value{
A list containing MAJIQ workflow.
}
\description{
Generating commands of MAJIQ workflow.
}
\references{
\url{https://majiq.biociphers.org}.
}
