% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{plot_asp}
\alias{plot_asp}
\title{Plot tracks of bam files to visualise significant differential AS genes}
\usage{
plot_asp(
  asp,
  tool,
  plot_samples = NULL,
  dat = NULL,
  dir_plot = NULL,
  top_n = 10,
  ...
)
}
\arguments{
\item{asp}{An \code{\linkS4class{ASP}} object.}

\item{tool}{Any combination of "rmats", "cash", "leafcutter", "majiq",
"suppa", "bandits", "psichomics", and "spladder".}

\item{plot_samples}{Can be a integer (sampling in each condition)
or a character vector (sample names).}

\item{dat}{If it's NULL, plot \code{top_n}.}

\item{dir_plot}{If it's NULL, pdf files will be deposited under the subfolder
\emph{track} of \emph{dir_out} stored in \code{\linkS4class{ASP}}.}

\item{top_n}{Default is 10.}

\item{...}{Accept two additional flags: \strong{extend_left} and
\strong{extend_right}, mean how many based should be extended around the
AS events.}
}
\value{
An invisible \code{\linkS4class{ASP}} object.
}
\description{
This function is to plot tracks
    using \code{\link[trackViewer]{trackViewer-package}}.
}
