% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{plot_pre_asp}
\alias{plot_pre_asp}
\title{Prepare for plotting tracks of significant differential AS genes}
\usage{
plot_pre_asp(asp, gtf = NULL, annotation_db = NULL)
}
\arguments{
\item{asp}{An \code{\linkS4class{ASP}} object.}

\item{gtf}{If it's NULL, the gtf stored in \code{\linkS4class{ASP}} will be
used.}

\item{annotation_db}{OrgDb.}
}
\value{
An \code{\linkS4class{ASP}} object.
}
\description{
This function is to load a txdb and an annotation database.
}
