% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{psichomics}
\alias{psichomics}
\title{psichomics}
\usage{
psichomics(
  dir_out,
  basics,
  sampletable,
  conda_path,
  conda_env,
  novel,
  write_log,
  anno_suppa = "",
  anno_rmats = "",
  anno_miso = "",
  anno_vast = "",
  prefix_suppa = "",
  prefix_miso = "",
  prefix_vast = "",
  types_miso = "",
  libpaths = "",
  ...
)
}
\arguments{
\item{dir_out, basics, sampletable, novel, conda_path, conda_env, write_log}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{anno_suppa}{The directory holding SUPPA event files.}

\item{anno_rmats}{The directory holding rMATS event files.}

\item{anno_miso}{The directory holding MISO event files.}

\item{anno_vast}{The directory holding VAST-TOOLS event files.}

\item{prefix_suppa}{No default. The prefix of SUPPA event files.}

\item{prefix_miso}{No default. The prefix of MISO event files.}

\item{prefix_vast}{No default. The prefix of VAST-TOOLS event files.}

\item{types_miso}{Should be like this: c("SE", "MXE", "A5SS", "A3SS", "RI").}

\item{libpaths}{Default is "". If not NULL, should be like this:
c("~/doc/Rlibrary/", "~/doc/Rlibrary4/")}

\item{...}{nothing}
}
\value{
A list containing psichomics workflow.
}
\description{
Generating commands of psichomics workflow.
}
\references{
\url{https://bioconductor.org/packages/release/bioc/html/psichomics.html}.
}
