% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{rmats}
\alias{rmats}
\title{rMATS}
\usage{
rmats(
  dir_out,
  basics,
  nproc,
  gtf,
  novel,
  conda_path,
  conda_env,
  write_log,
  variable_read_length = F,
  paired_stats = F,
  cstat = 1e-04,
  allow_clipping = T,
  ...
)
}
\arguments{
\item{dir_out, basics, nproc, gtf, novel, conda_path, conda_env, write_log}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{variable_read_length}{Default is FALSE.}

\item{paired_stats}{Default is FALSE.}

\item{cstat}{Default is 0.0001.}

\item{allow_clipping}{Default is TRUE.}

\item{...}{nothing}
}
\value{
A list containing rMATS workflow.
}
\description{
Generating commands of rMATS workflow.
}
\references{
\url{https://github.com/Xinglab/rmats-turbo}.
}
