% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{run_asp}
\alias{run_asp}
\title{Run the command generated by \code{\link{cmd_asp}} and/or
    read in the results}
\usage{
run_asp(asp, tool, run = T, read_results = T, parallel = T, block = T)
}
\arguments{
\item{asp}{An \code{\linkS4class{ASP}} object.}

\item{tool}{Any combination of "rmats", "cash", "leafcutter", "majiq",
"suppa", "bandits", "psichomics", and "spladder".}

\item{run}{Default is TRUE. If you just want to check the command, set to
FALSE, and the whole workflow will be printed.}

\item{read_results}{Default is TRUE.
If read in results after workflows are done.}

\item{parallel}{Default is TRUE. If run multiple workflows simultaneously.
The function \code{\link[future]{future}} is imposed.}

\item{block}{Default is TRUE. If wait for the end of running.}
}
\value{
An \code{\linkS4class{ASP}} object with results saved in results slot.
}
\description{
This function is about to run each workflow and read in results.
}
\examples{
\dontrun{
asp <- run_asp(
  asp = asp, tool = "rmats", run = T, read_results = T, parallel = F
)
asp <- run_asp(
  asp = asp, tool = c("cash", "leafcutter"),
  read_results = T, parallel = T, block = T
)
}
}
