% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{sig_asp}
\alias{sig_asp}
\title{Query significant results}
\usage{
sig_asp(
  asp,
  tool,
  p_adj = 0.05,
  p_value = NULL,
  delta_psi = 0,
  statistic_col = NULL,
  statistic_threshold = 0,
  filter_direction = c(">", "<"),
  effect_size_col = NULL,
  effect_size_threshod = 0
)
}
\arguments{
\item{tool}{Any combination of "rmats", "cash", "leafcutter", "majiq",
"suppa", "bandits", "psichomics", and "spladder".}

\item{p_adj}{The threshold for the padj column.}

\item{p_value}{The threshold for the p column.}

\item{delta_psi}{The threshold for the dpsi column.}

\item{statistic_col}{If results have no padj or p column, provide an
alternative.}

\item{statistic_threshold}{The threshold for the alternative
statistics column.}

\item{filter_direction}{One of \code{c(">", "<")} indicating the significance
direction.}

\item{effect_size_col}{If results have no dpsi column, provide an
alternative.}

\item{effect_size_threshod}{The threshold for the alternative
effect size column.}
}
\value{
An \code{\linkS4class{ASP}} object with significant results saved
    in sig_results slot.
}
\description{
This function is about to extract the significant results.
}
