% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{spladder}
\alias{spladder}
\title{SplAdder}
\usage{
spladder(
  dir_out,
  basics,
  gtf,
  novel,
  novel_ss,
  conda_path,
  conda_env,
  write_log,
  parallel,
  np,
  confidence_level = 3,
  ...
)
}
\arguments{
\item{dir_out, basics, gtf, novel, novel_ss, parallel, np, conda_path, conda_env, write_log}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{confidence_level}{Default is 3.}

\item{...}{nothing}
}
\value{
A list containing SplAdder workflow.
}
\description{
Generating commands of SplAdder workflow.
}
\references{
\url{https://spladder.readthedocs.io/en/latest/}.
}
