% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{suppa}
\alias{suppa}
\title{SUPPA}
\usage{
suppa(
  dir_out,
  gtf,
  np,
  conda_path,
  conda_env,
  parallel,
  write_log,
  dir_events = NULL,
  ...
)
}
\arguments{
\item{dir_out, np, gtf, conda_path, conda_env, parallel, write_log}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{...}{nothing}

\item{dir_envents}{Any directory is ok.
The directory holding SUPPA ioe files. Keep NULL will make the ioe files
be generated in the \emph{dir_out/SUPPA/events}.}
}
\value{
A list containing SUPPA workflow.
}
\description{
Generating commands of SUPPA workflow.
}
\references{
\url{https://github.com/comprna/SUPPA}.
}
