% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tx2gene}
\alias{tx2gene}
\title{A tibble containing chromosome, gene ensemble id, gene symbol and transcript
    ensemble id for each transcript documented by \emph{GENCODE version32}.}
\format{
A tibble with 227462 rows and 4 variables:
\describe{
    \item{chr}{chromosome name}
    \item{gene_id}{gene ensemble ID}
    \item{gene_symbol}{gene symbol}
    \item{transcript_id}{transcript ensemble ID}
}
}
\usage{
tx2gene
}
\description{
A tibble containing chromosome, gene ensemble id, gene symbol and transcript
    ensemble id for each transcript documented by \emph{GENCODE version32}.
}
\keyword{datasets}
