% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{venn_upset_asp}
\alias{venn_upset_asp}
\title{Visualize the intersection}
\usage{
venn_upset_asp(asp, tools = "all", type = "all")
}
\arguments{
\item{asp}{An \code{\linkS4class{ASP}} object.}

\item{type}{One of "SE", "MXE", "A3SS", "A5SS", "RI", "SME", "ALE" and "AFE".
Or use "all" to use the intersections of all AS types.}

\item{tool}{Any combination of "rmats", "cash", "leafcutter", "majiq",
"suppa", "bandits", "psichomics", and "spladder".
Or use "all" to ues the intersections of all significant results
stored in your \code{\linkS4class{ASP}} object.}
}
\value{
A ggplot object.
}
\description{
This function is to draw upset and venn plots.
}
\examples{
\dontrun{
venn_upset_asp(asp)
}
}
