% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ASP-class}
\alias{ASP-class}
\alias{ASP}
\title{Class \code{ASP}}
\description{
All information stored in ASP object.
   You can use \code{creatASP} to create an ASP object.
   In this package, most of the functions will use
   ASP object as input, and return a modified ASP as well.
}
\section{Slots}{

\describe{
\item{\code{sampletable}}{data.frame. An experiment design table and
these columns are mandatory:
\itemize{
    \item samples
    \item conditions
    \item read_lengths
    \item library_types
    \item strandedness
}
, and these columns are dispensable but at least 1 needs to be provided:
\itemize{
    \item files_bam
    \item files_fq
    \item dirs_salmon
}}

\item{\code{tx2gene}}{data.frame. At least three types of
information for each transcript should be provided:
\itemize{
    \item gene_id
    \item gene_symbol
    \item transcript_id
}
If not provided, automatically generated from the gtf slot.}

\item{\code{tx2gene_convert}}{data.frame. Automatically generated from
the tx2gene slot.}

\item{\code{gtf}}{character. The path to your gtf file.}

\item{\code{gff}}{character. The path to your gff3 file.}

\item{\code{fa}}{character. The path to your genome fasta file.}

\item{\code{dir_out}}{character. The parent directory for all the next procedures.
Default is "./".}

\item{\code{nproc}}{integer. The number of threads. Default is 5.}

\item{\code{novel}}{boolean. If to detect novel splicing or not. Default is FALSE.}

\item{\code{novel_ss}}{boolean. If to detect novel splice site or not. Default is FALSE.}

\item{\code{write_log}}{boolean. If to write the stdout and stderr to a log file
beginning with an underscore. Default is TRUE.}

\item{\code{parallel}}{boolean. If to run different tools or multiple samples
in parallel. Default is TRUE.}

\item{\code{np}}{integer. How many samples will be processed at the same time.
Default is 5.}

\item{\code{conda_path}}{character. The path to your conda or vritual environment.
Different tools can rely on different paths.}

\item{\code{basics}}{list. Automatically generated based on the sampletable.}

\item{\code{cmds}}{list. Storing commands of different tools to run or to check
generated by \code{\link{cmd_asp}}.}

\item{\code{results}}{list. Storing results outputted by different tools
read by \code{\link{run_asp}}.}

\item{\code{sig_results}}{list. Results fulfilling significant thresholds
generated by \code{\link{sig_asp}}.}

\item{\code{parameters}}{list.}

\item{\code{intersection}}{list. Results intersections generated by
\code{\link{intersect_asp}}.}

\item{\code{txdb}}{TxDb. Can be generated automatically from the gtf slot by
\code{\link[GenomicFeatures]{makeTxDbFromGFF}}.}

\item{\code{annotation_db}}{OrgDb. Bioconductor annotation data package.
For example: \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}.}

\item{\code{bams}}{trackList. \code{\link[trackViewer]{trackList-class}}.
Storing GRanges of bams you want to visualize.}

\item{\code{ranges_lines}}{list. Storing GRanges, lines, IDs, and genes of the
significant differential alternative events.}

\item{\code{basics_plot}}{list. Storing indications of two/one group(s) of bams.}
}}

