% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{cmd_asp}
\alias{cmd_asp}
\title{Generate commands for AS tools}
\usage{
cmd_asp(
  asp,
  conda_env = NULL,
  tool,
  tx2gene = NULL,
  dir_out = NULL,
  nproc = NULL,
  gtf = NULL,
  gff = NULL,
  fa = NULL,
  parallel = NULL,
  np = NULL,
  basics = NULL,
  sampletable = NULL,
  novel = NULL,
  conda_path = NULL,
  write_log = NULL,
  ...
)
}
\arguments{
\item{asp}{An \code{\linkS4class{ASP}} object.}

\item{conda_env}{The path to environment each tools depend on.
Different tools can rely on different envs.}

\item{tool}{Any combination of "rmats", "cash", "leafcutter", "majiq",
"suppa", "bandits", "psichomics", and "spladder".}

\item{tx2gene, dir_out, nproc, gtf, gff, fa}{These are same as the flags in
\code{\linkS4class{ASP}}, if you want to change any of them, then
provide a new one.}

\item{parallel, np, basics, sampletable, novel, conda_path, write_log}{These are same
as the flags in \code{\linkS4class{ASP}}, if you want to change any of
them, then provide a new one.}

\item{...}{paramters pass to \code{\link{rmats}}, \code{\link{cash}},
\code{\link{leafcutter}}, \code{\link{majiq}}, \code{\link{suppa}},
\code{\link{bandits}}, \code{\link{psichomics}},
and \code{\link{spladder}}.}
}
\value{
An \code{\linkS4class{ASP}} object with commands of each workflows
    saved in cmds slot.
}
\description{
This function is about to generate desire commands to run or to
    check.
}
\examples{
\dontrun{
asp <- cmd_asp(
  asp = asp, conda_env = "rmats", tool = "rmats",
  variable_read_length = T, paired_stats = F, cstat = 0.01, allow_clipping = T
)

asp <- cmd_asp(
  asp = asp, conda_env = c(NA, "assw_3.8"), tool = c("cash", "leafcutter"),
  cash_jar = "~/bin/cash.jar", MergePval = "G", ram = "100g",
  leafcutter_dir = "~/bin/leafcutter/"
)
}
}
