% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{intersect_asp}
\alias{intersect_asp}
\title{Intersect the results}
\usage{
intersect_asp(asp, tools = "all", type = "all")
}
\arguments{
\item{asp}{An \code{\linkS4class{ASP}} object.}

\item{tools}{Any combination of "rmats", "cash", "leafcutter", "majiq",
"suppa", "bandits", "psichomics", and "spladder".
Or use "all" to retrieve the intersections of all significant results
stored in your \code{\linkS4class{ASP}} object.}

\item{type}{One of "SE", "MXE", "A3SS", "A5SS", "RI", "SME", "ALE" and "AFE".
Or use "all" to retrieve the intersections of all AS types.}
}
\value{
An \code{\linkS4class{ASP}} object with intersection tibble
    stored in the intersection slot.
}
\description{
This function is to retrieve intersections of significant
    results of different workflows.
}
\examples{
\dontrun{
asp <- intersect_asp(asp)
print(asp@intersection$all)
}
}
