% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{plot_asp}
\alias{plot_asp}
\title{Plot tracks of bam files to visualise significant differential AS genes}
\usage{
plot_asp(asp, tool, i = 1)
}
\arguments{
\item{asp}{An \code{\linkS4class{ASP}} object.}

\item{tool}{One of "rmats", "cash", "leafcutter", "majiq",
"suppa", "psichomics", or "spladder".}

\item{i}{Default is 1. The order of the significant differential events
prepared through \code{\link{plot_pre_asp}} you want to plot.}
}
\value{
NULL.
}
\description{
This function is to plot tracks
    using \code{\link[trackViewer]{trackViewer-package}}.
}
\examples{
\dontrun{
plot_asp(asp, tool = "cash", i = 1)
}
}
