% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{psichomics}
\alias{psichomics}
\title{psichomics}
\usage{
psichomics(
  dir_out,
  basics,
  sampletable,
  conda_path,
  conda_env,
  novel,
  write_log,
  anno_suppa = NULL,
  anno_rmats = NULL,
  anno_miso = NULL,
  anno_vast = NULL,
  prefix_suppa = NULL,
  prefix_miso = NULL,
  prefix_vast = NULL,
  types_miso = NULL,
  ...
)
}
\arguments{
\item{dir_out, basics, sampletable, novel, conda_path, conda_env, write_log}{These are obtained from \code{\linkS4class{ASP}} object.}

\item{anno_suppa}{The directory holding SUPPA event files.}

\item{anno_rmats}{The directory holding rMATS event files.}

\item{anno_miso}{The directory holding MISO event files.}

\item{anno_vast}{The directory holding VAST-TOOLS event files.}

\item{prefix_suppa}{No default. The prefix of SUPPA event files.}

\item{prefix_miso}{No default. The prefix of MISO event files.}

\item{prefix_vast}{No default. The prefix of VAST-TOOLS event files.}

\item{types_miso}{Should be like this: "SE,MXE,A5SS,A3SS,RI".}

\item{...}{nothing}
}
\value{
A list containing psichomics workflow.
}
\description{
Generating commands of psichomics workflow.
}
\references{
\url{https://bioconductor.org/packages/release/bioc/html/psichomics.html}.
}
